review: move copy_buffer.rs code to copy_bidirectional.rs
